+++ client/src/main/java/com/showmeyourcode/spring_cloud/client/ClientApplication.java
@@ -9,9 +9,9 @@ import org.springframework.cloud.openfeign.EnableFeignClients;
 		"com.showmeyourcode.spring_cloud.client"
 })
 @SpringBootApplication
-public class SpringBootCloudClientApplication {
+public class ClientApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(SpringBootCloudClientApplication.class, args);
+		SpringApplication.run(ClientApplication.class, args);
 	}
 }