+++ client/src/main/java/com/showmeyourcode/spring_cloud/client/configuration/SwaggerConfig.java
@@ -6,17 +6,17 @@ import springfox.documentation.builders.PathSelectors;
 import springfox.documentation.builders.RequestHandlerSelectors;
 import springfox.documentation.spi.DocumentationType;
 import springfox.documentation.spring.web.plugins.Docket;
-import springfox.documentation.swagger2.annotations.EnableSwagger2WebMvc;
+import springfox.documentation.swagger2.annotations.EnableSwagger2;
 
 @Configuration
-@EnableSwagger2WebMvc
+@EnableSwagger2
 public class SwaggerConfig {
 
     @Bean
     public Docket api() {
         return new Docket(DocumentationType.OAS_30)
                 .select()
-                .apis(RequestHandlerSelectors.basePackage("com.showmeyourcode.spring_cloud.client.rest"))
+                .apis(RequestHandlerSelectors.basePackage("com.showmeyourcode.spring_cloud.client.api"))
                 .paths(PathSelectors.any())
                 .build();
     }