+++ eureka-server/src/test/java/com/showmeyourcode/spring_cloud/eureka_server/EurekaApplicationTest.java
@@ -8,7 +8,7 @@ import org.springframework.context.ApplicationContext;
 import static org.assertj.core.api.Assertions.assertThat;
 
 @SpringBootTest
-class SpringBootCloudEurekaApplicationTest {
+class EurekaApplicationTest {
 
     @Autowired
     private ApplicationContext context;