+++ microservice1/src/main/java/com/showmeyourcode/spring_cloud/microservice1/Microservice1Application.java
@@ -4,9 +4,9 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class SpringBootCloudMicroservice1Application {
+public class Microservice1Application {
 
 	public static void main(String[] args) {
-		SpringApplication.run(SpringBootCloudMicroservice1Application.class, args);
+		SpringApplication.run(Microservice1Application.class, args);
 	}
 }