+++ microservice1/src/main/java/com/showmeyourcode/spring_cloud/microservice1/api/Microservice1Endpoint.java
@@ -4,7 +4,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 
 public interface Microservice1Endpoint {
 
-    String ENDPOINT_PATH = "/api/microservice1";
+    String ENDPOINT_PATH = "/api/v1";
 
     @RequestMapping(ENDPOINT_PATH)
     String getMicroserviceName();