+++ microservice2/src/main/java/com/showmeyourcode/spring_cloud/microservice2/api/Microservice2Endpoint.java
@@ -4,7 +4,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 
 public interface Microservice2Endpoint {
 
-    String ENDPOINT_PATH = "/api/microservice2";
+    String ENDPOINT_PATH = "/api/v1";
 
     @RequestMapping(ENDPOINT_PATH)
     String getMicroserviceName();