+++ CHANGELOG.md
@@ -4,12 +4,29 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2.0.0 (????-??-??)
+
+### Added
+
+- ???
+
+### Changed
+
+- API specification for microservices ([xx](xx))
+  - Add versioning and context paths
+
+### Fixed
+
+- ???
+
+###
+
 ## 1.0.0 (2023-06-14)
 
 ### Added
 
-- Spring Cloud services (Eureka Server, Eureka clients Ribbon)
+- Spring Cloud services (Eureka Server, OpenFeign, Client Side Load Balancer: Ribbon)
 - Spring Boot Admin
-- Swagger
-- Spring REST Docs
+- Swagger API for all microservices
+- Spring REST Docs for the client microservice
 - Gitlab CI/CD configuration